# [R] Speed-up lgb.importance()