--- a/R-package/R/lgb.model.dt.tree.R
+++ b/R-package/R/lgb.model.dt.tree.R
@@ -90,14 +90,25 @@ lgb.model.dt.tree <- function(model, num_iteration = NULL) {
 
 #' @importFrom data.table := data.table rbindlist
 .single_tree_parse <- function(lgb_tree) {
-
+  long_list <- c(
+    "split_index"
+    , "split_feature"
+    , "split_gain"
+    , "threshold"
+    , "decision_type"
+    , "default_left"
+    , "internal_value"
+    , "internal_count"
+  )
+  
   # Traverse tree function
   pre_order_traversal <- function(env = NULL, tree_node_leaf, current_depth = 0L, parent_index = NA_integer_) {
-
+    
     if (is.null(env)) {
       # Setup initial default data.table with default types
       env <- new.env(parent = emptyenv())
-      env$single_tree_dt <- data.table::data.table(
+      env$single_tree_dt <- list()
+      env$single_tree_dt[[1L]] <- data.table::data.table(
         tree_index = integer(0L)
         , depth = integer(0L)
         , split_index = integer(0L)
@@ -122,25 +133,16 @@ lgb.model.dt.tree <- function(model, num_iteration = NULL) {
         , parent_index = parent_index
       )
     } else {
-
+      
       # Check if split index is not null in leaf
       if (!is.null(tree_node_leaf$split_index)) {
-
+        
         # update data.table
-        env$single_tree_dt <- data.table::rbindlist(l = list(env$single_tree_dt,
-                                                             c(tree_node_leaf[c("split_index",
-                                                                                "split_feature",
-                                                                                "split_gain",
-                                                                                "threshold",
-                                                                                "decision_type",
-                                                                                "default_left",
-                                                                                "internal_value",
-                                                                                "internal_count")],
-                                                               "depth" = current_depth,
-                                                               "node_parent" = parent_index)),
-                                                    use.names = TRUE,
-                                                    fill = TRUE)
-
+        env$single_tree_dt[[length(env$single_tree_dt) + 1L]] <- c(
+          tree_node_leaf[long_list]
+          , list("depth" = current_depth, "node_parent" = parent_index)
+        )
+        
         # Traverse tree again both left and right
         pre_order_traversal(
           env = env
@@ -154,31 +156,29 @@ lgb.model.dt.tree <- function(model, num_iteration = NULL) {
           , current_depth = current_depth + 1L
           , parent_index = tree_node_leaf$split_index
         )
-
+        
       } else if (!is.null(tree_node_leaf$leaf_index)) {
-
-        # update data.table
-        env$single_tree_dt <- data.table::rbindlist(l = list(env$single_tree_dt,
-                                                             c(tree_node_leaf[c("leaf_index",
-                                                                                "leaf_value",
-                                                                                "leaf_count")],
-                                                               "depth" = current_depth,
-                                                               "leaf_parent" = parent_index)),
-                                                    use.names = TRUE,
-                                                    fill = TRUE)
-
+        
+        # update list
+        env$single_tree_dt[[length(env$single_tree_dt) + 1L]] <- c(
+          tree_node_leaf[c("leaf_index", "leaf_value", "leaf_count")]
+          , list("depth" = current_depth, "leaf_parent" = parent_index)
+        )
       }
-
     }
     return(env$single_tree_dt)
   }
-
-  # Traverse structure
-  single_tree_dt <- pre_order_traversal(tree_node_leaf = lgb_tree$tree_structure)
-
+  
+  # Traverse structure and rowbind everything
+  single_tree_dt <- data.table::rbindlist(
+    pre_order_traversal(tree_node_leaf = lgb_tree$tree_structure)
+    , use.names = TRUE
+    , fill = TRUE
+  )
+  
   # Store index
   single_tree_dt[, tree_index := lgb_tree$tree_index]
-
+  
   return(single_tree_dt)
-
 }
+