`lgb.importance()` is relatively slow, partly due to 

1. the json writing/reading, and
2. recursive calls to `data.table::rbindlist()`

This PR modifies the `lgb.model.dt.tree()` workhorse as follows: instead of rbinding every single node, the node information is added to a (growing) list. At the end, a single call to `rbindlist()` combines all node infos.

## Speed-up

This demo with 500 trees and 50k training data, roughly reduces the time from 5.5s to 2s. The results are unchanged.
```
library(lightgbm)
library(ggplot2)
library(bench)
library(profvis)

params <- list(
  objective = "regression"
  , metric = "l2"
  , learning_rate = 0.05
)
model <- lgb.train(
  params = params
  , data = lgb.Dataset(data.matrix(diamonds[1:4]), label = diamonds$price)
  , nrounds = 500
)

system.time(  # 5.5s -> 2.0s
  imp <- lgb.importance(model)
)
imp

# Feature        Gain      Cover Frequency
# <char>       <num>      <num>     <num>
# 1:   carat 0.893368734 0.75091164 0.5106000
# 2: clarity 0.068601335 0.10038648 0.1738000
# 3:   color 0.034605234 0.09241368 0.1857333
# 4:     cut 0.003424697 0.05628819 0.1298667

profvis(lgb.importance(model))
```

## Profiler

In above example, about 2/3 of the time is spent in json parsing. Thus, the calculation part is reduced by a factor of about 6.

![image](https://github.com/microsoft/LightGBM/assets/16206095/e705bfaa-734f-4a9c-b076-c6b66666cea9)