@jameslamb 2024-03-16
Excellent, thank you for the great work and thorough write-up!!!

Many of the CI failures look unrelated to these changes, I will try to help with them (and provide a review here) soon.