@mayer79 2024-03-18
A couple of unit tests would certainly make sense. I will try to add some.