@mayer79 2024-03-18
@jameslamb Only partly what you suggested: regression only, and no checks for `lgb.importance()`. I can add more tests later.