@jameslamb 2024-03-19
@mayer79 the tests you added look great, thanks!

Do you think you'll have time to add more in the next few days? If not, I think we should merge this as-is and then just document the additional testing work in a separate issue.