@mayer79 2024-03-19
Great! I will add a loop over some more models fitted with different objectives.

Maybe we can add tests for `lgb.importance()` in a later PR? I can help with these too (next week).