@mayer79 2024-03-19
@jameslamb added a test for `num_iterations` and a loop over three model types.