@jameslamb 2024-04-10
Thanks for your help and patience @mayer79 !